requirements.txt
Django==4.2
django-model-utils~=4.3.1
django-bootstrap-v5~=1.0.11